test pkg deleted, main package renamed, added colors for landscape
