+++ app/src/main/res/layout-land/activity_main.xml
@@ -1,20 +1,17 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout
+<LinearLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:background="@color/lightblue"
-        android:gravity="center"
-        android:orientation="vertical">
-
+    tools:context=".MainActivity"
+    android:background="@color/mainbkg"
+    android:gravity="center"
+    android:orientation="vertical"
+    >
         <TextView
-            android:id="@+id/tv"
+            android:id="@+id/mainGreetingsTv"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="@string/msg"
@@ -42,4 +39,3 @@
             android:paddingLeft="10dp"/>
 
     </LinearLayout>
-  </androidx.constraintlayout.widget.ConstraintLayout>
