+++ app/src/main/res/layout/activity_main.xml
@@ -1,23 +1,21 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout
+<LinearLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="vertical"
-        android:gravity="center"
-        android:background="@color/teal_200"
-        android:id="@+id/LL"
-        >
+    tools:context="com.sv053.reddit.MainActivity"
+    android:gravity="center"
+    android:background="@color/mainbkg"
+    android:id="@+id/LL"
+    android:orientation="vertical"
+    >
+
         <TextView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:id="@+id/tv"
+            android:id="@+id/mainGreetingsTv"
             android:text="@string/msg"
             android:textAlignment="center"
             android:textSize="28sp"
@@ -42,4 +40,3 @@
             android:paddingLeft="10dp"/>
 
     </LinearLayout>
-  </androidx.constraintlayout.widget.ConstraintLayout>
