+++ app/src/main/res/values/colors.xml
@@ -8,4 +8,5 @@
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
     <color name="lightblue">#B4EDFB</color>
+    <color name="mainbkg">#FF03DAC5</color>
 </resources>
